+++ covidapi/api/views.py
@@ -3,8 +3,9 @@ from fastapi import APIRouter
 from sqlalchemy.orm import Session
 from typing import List
 
-from ..db import crud, models
-from ..db.schemas import DailyReport
+from ..db import models
+from ..services import crud
+from ..schemas.schemas import DailyReport
 from ..db.database import SessionLocal, engine
 
 models.Base.metadata.create_all(bind=engine)
