+++ covidapi/schemas/schemas.py
